made an API change to enforce the right aura methods, also add over- and underflow protection

+++ src/main/java/de/ellpeck/naturesaura/events/ClientEvents.java
@@ -57,9 +57,9 @@ public class ClientEvents {
                 MutableInt spots = new MutableInt();
                 IAuraChunk.getSpotsInArea(mc.world, mc.player.getPosition(), 15, ((blockPos, drainSpot) -> {
                     spots.increment();
-                    amount.add(drainSpot.intValue());
+                    amount.add(drainSpot);
                     if (mc.player.isSneaking())
-                        left.add(prefix + drainSpot.intValue() + " @ " + blockPos.getX() + " " + blockPos.getY() + " " + blockPos.getZ());
+                        left.add(prefix + drainSpot + " @ " + blockPos.getX() + " " + blockPos.getY() + " " + blockPos.getZ());
                 }));
                 left.add(prefix + "Total: " + amount.intValue() + " in " + spots.intValue() + " spots");
                 left.add(prefix + "Type: " + IAuraType.forWorld(mc.world).getName());
